pattern: allow disabling `Class` to support generalized patterns
Because type bounds propagate from `Class` to `Element`, we must
fully excise `Class` if we wish to use patterns with non-trivial types
for which there is no logical way to implement the `Step` and `Distance`
traits.

To prevent any loss of expressiveness when converting from
`regex_syntax` regular expressions, `Union` now implements `From`
for that crate's Class types.