pattern: avoid exposing `regex_syntax` dependency to client crates
By introducing a `TryFrom<&str>` impl for element types that also
implement `TryFrom<regex_syntax::Expr>`, we can prevent our
dependency on `regex_syntax` from leaking through during conversion.
Previously the type bounds on `parse_regex` explicitly exposed this
dependency.

There is, of course, still a potential issue with the
`regex_syntax::Error` instance in a variant of `RegexConvResult`; we'll
deal with that when it becomes an issue.